FInal changes -> fixed x86 bug in reader

+++ Minigin/BinaryReader.cpp
@@ -8,7 +8,7 @@ std::string BinaryReader::ReadString()
 	if (m_InputStream.is_open())
 	{
 		//We read in the size, resize the string and read in the characters
-		unsigned int stringSize{};
+		size_t stringSize{};
 		m_InputStream.read((char*)& stringSize, sizeof(stringSize));
 		toRead.resize(stringSize);
 		m_InputStream.read((char*)& toRead[0], toRead.size());
@@ -19,7 +19,6 @@ std::string BinaryReader::ReadString()
 void BinaryReader::Open()
 {
 	m_InputStream.open(m_Filename, std::ios::in | std::ios::binary);
-
 }
 
 void BinaryReader::Close()
